Wumpex.Api.Ratelimit (wumpex v0.1.0)

The Wumpex.Api.Ratelimit module handles dispatching HTTP requests through the Wumpex.Api module, but handles rate limits.

Examples

iex> Wumpex.Api.Ratelimit.get("http://localhost/test")
{:ok, _response}

Types

bucket_tag()

Specs

bucket_tag() :: tuple()

state()

Specs

state() :: %{
  buckets: :ets.tid(),
  bucket_states: :ets.tid(),
  well: GenServer.server()
}

The state of the Wumpex.Api.Ratelimit process.

Contains the following values:

:buckets - The :ets table containing all the bucket identifiers

:bucket_states - The :ets table containing the state for all buckets.

:well - The GenServer.server/0 of the Wumpex.Api.Ratelimit.Well process.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

request(http_call, tag, timeout \\ 5000)

Specs

request(
  Wumpex.Api.Ratelimit.StatelessBucket.http_call(),
  bucket_tag(),
  timeout()
) :: Wumpex.Api.Ratelimit.StatelessBucket.http_response() | {:error, :bounced}

Executes a given request using Wumpex.Api.request/5 using ratelimits.

This means that if a rate limit would be encountered the request will either be bounced or delayed, depending on the configured timeout and when the resource will become available.

For example, imagine the URL "/test" can be called once per 10 seconds:

iex> Wumpex.Api.Ratelimit.request({:get, "http://localhost/test", "", [], []}, {"ratelimit", "demo"})
  {:ok, %HTTPoison.Response{}}
iex> Wumpex.Api.Ratelimit.request({:get, "http://localhost/test", "", [], []}, {"ratelimit", "demo"})
  {:error, :bounced}

The first request executes and returns the response (an instance of HTTPoison.Response.t/0), but the second request gets {:error, :bounced}. This is because after the first request we'd have to wait 10s, while we specified we only want to wait up to 5s (default, can be overriden by passing in a third parameter). If we'd have passed in 11s for example, the request would have been delayed 10s before executing, after which the response would be returned.

You can choose to always execute the request, regardless of how long it would take until it can execute by passing :infinity as the third parameter.