Wumpex.Api.Ratelimit.StatelessBucket (wumpex v0.1.0)

Represents a single "bucket" in the rate limit system of Discord as described in the official documentation.

Buckets are responsible for checking whether the given request is allowed to execute, and if not queue or bounce the request according to the given configuration. The StatelessBucket module gets an ETS table passed in on every request that contains the state for that specific bucket. This allows executing multiple requests on the same Discord bucket (but spread across multiple processes) and having a worker pool of buckets (using :poolboy).

Types

bucket_state()

Specs

bucket_state() :: %{
  remaining: non_neg_integer() | nil,
  reset_at: non_neg_integer() | nil
}

The state stored in the :ets table passed in command/0.

command()

Specs

command() :: %{
  http: http_call(),
  timeout: timeout(),
  state: :ets.tid(),
  bucket: any()
}

Represents a command for the bucket to execute.

Each command contains 4 fields:

:http - A http_call/0 that represents the HTTP call to execute (if possible).

:timeout - A timeout/0 that indicates how long the caller is willing to wait in case we need to queue.

:state - An :ets.tid/0 table that contains the state for the http call resource (eg. remaining rate limit, reset, ...).

:bucket - The name of the bucket in which this command will execute (used for state lookup).

http_call()

Specs

http_call() ::
  {HTTPoison.Request.method(), HTTPoison.Request.url(),
   HTTPoison.Request.body(), HTTPoison.Request.headers(),
   HTTPoison.Request.options()}

Contains all the information required for making a HTTP call.

The elements correspond to the parameters of Wumpex.Api.request/5.

http_response()

Specs

http_response() ::
  {:ok, HTTPoison.Response.t()}
  | {:error, HTTPoison.Error.t() | HTTPoison.Response.t()}

Represents the result of an executed http_call/0.

state()

Specs

state() :: nil

This module does not keep any state in the process.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

execute(http_call)

Specs

execute(http_call()) :: http_response()

Executes the given http_call/0 and returns the http_response/0.

Calls are executed using Wumpex.Api.request/5.