Wumpex.Api.Ratelimit.Well (wumpex v0.1.0)

The well handles requests for routes which Wumpex doesn't know the bucket for.

It executes the incoming requests, and then attempts to associate the key for that route to a bucket.

Types

command()

Specs

command() :: %{
  http: Wumpex.Api.Ratelimit.StatelessBucket.http_call(),
  tag: Wumpex.Api.Ratelimit.bucket_tag(),
  buckets: :ets.tid(),
  bucket_states: :ets.tid()
}

state()

Specs

state() :: nil

This module does not keep any state in the process.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.