Wumpex.Base.Websocket behaviour (wumpex v0.1.0)

A behaviour module for implementing the client of a websocket connection.

A Websocket is a process like a common GenServer but handles the opening of the websocket transparently for you. All extensions made to the regular GenServer have been designed to closely match the behaviour of a regular GenServer.

For example the handle_frame callback has very similiar return values to the handle_call callback of the GenServer.

Options

The Websocket requires a few options to be passed in the start_link method:

:host - the hostname (ex: "gateway.discord.gg").

:port - the port to connect to (usually 80 for unencrypted and 443 for encrypted connections).

:path - the path to which the websocket will connect (ex: "/?v=8") and can (but doesn't have to) contain a query string.

:timeout is a timeout/0 used to determine how long to wait for connecting and upgrading.

See options/0 for more information.

Example

A module implementing the behaviour will be able to respond to incoming messages by implementing handle_frame:

defmodule ExampleWebsocket do
  use Wumpex.Base.Websocket

  require Logger

  def on_connected(_options) do
    Logger.info("Connected!")
  end

  def handle_frame({:text, message}, state) do
    Logger.info("Received message")

    {:noreply, state}
  end
end

Types

frame()

Specs

frame() ::
  {:text, String.t()}
  | {:binary, iodata()}
  | {:close, non_neg_integer(), iodata()}
  | :close

Represents a single websocket frame that can be sent or received over the websocket.

There's three supported frames:

{:text, message} - Used to send/receive textual data.

{:binary, message} - Used to send/receive binary data (like ETF).

:close or {:close, status, reason} - Signals that the websocket will be closed (to the receiving side), optionall you can pass in a status code and a reason.

:ping and :pong frames are handled internally and not exposed.

options()

Specs

options() :: [
  host: String.t(),
  port: non_neg_integer(),
  path: String.t(),
  timeout: timeout()
]

The options that are required to be passed to the websocket on startup.

host is the hostname (ex: "gateway.discord.gg"), port is the port to connect to (usually 80 for unencrypted and 443 for encrypted connections), path is the path to which the websocket will connect (ex: "/?v=8") and can (but doesn't have to) contain a query string and finally timeout is a timeout/0 used to determine how long to wait for connecting and upgrading.

All options are required to be passed in.

reconnect_strategy()

Specs

reconnect_strategy() :: :stop | :retry | {:retry, non_neg_integer()}

Reconnection strategy to use when returning from on_disconnected/1.

There are currently 3 reconnection strategies supported:

:stop - No reconnection, this will gracefully shut down the server.

:retry - Immediately attempt to reconnect.

{:retry, delay} - Wait for delay milliseconds, then attempt to reconnect.

Note that when using the delayed reconnect, the server will block and not process any messages until reconnection is complete. This is done to prevent receiving messages in the server when it is in a disconnected state.

Functions

send(websocket, frame_or_frames)

Specs

send(websocket :: pid(), frame_or_frames :: frame() | [frame()]) :: :ok

Send one or more frames over the given websocket.

websocket is the pid of the websocket, frame_or_frames is one (or a list of) frame/0.

start_link(module, init_args)

See GenServer.start_link/2.

start_link(module, init_args, options)

See GenServer.start_link/3.

Callbacks

handle_frame(frame, state)

Specs

handle_frame(frame :: frame(), state :: term()) ::
  {:reply, reply, new_state}
  | {:reply, reply, new_state, timeout()}
  | {:noreply, new_state}
  | {:noreply, new_state, timeout()}
  | {:stop, new_state}
  | {:stop, close_reason, new_state}
when reply: frame(),
     new_state: term(),
     close_reason: {non_neg_integer(), iodata()}

Invoked to handle incoming frames from the websocket connection.

frame is the frame/0 send by the server this websocket is connected to and state is the current state of the process.

Returning {:reply, frame, new_state} will dispatch frame on the websocket and set new_state as the state of the process, while returning {:noreply, new_state} will not send any frames and set new_state as the new state of the process.

You can also use the :reply and :noreply returns with an additional timeout/0 parameter, which behaves the same but will also set a timeout (See GenServer timeout section for more information).

You can also reply with {:stop, new_state} to gracefully close the connection (after sending a close frame to the server) and terminate the process, or {:stop, close_reason, new_state} to send a status and reason along in the close frame that's being sent to the server.

This callback and it's return types have intentionally been designed to be similiar to those of GenServer.handle_call/3, with the exception of the hibernate options, which is intentional.

on_connected(options)

Specs

on_connected(options :: keyword()) :: term()

Invoked when the websocket connects to the server.

This method is invoked the first time the websocket connects. This method can be compared with the init/1 method of the GenServer, and is used to set the state of the process.

It receives the options passed to start_link

The returned value will be used as the process state.

on_disconnected(state)
(optional)

Specs

on_disconnected(state :: term()) :: {term(), reconnect_strategy()}

Called when the socket connection closes.

state is the current process state, and the return value if this function is a tuple with the first element being the new process state and the second the reconnect_strategy/0 to use.

See reconnect_strategy/0 for more information.

on_reconnected(state)
(optional)

Specs

on_reconnected(state :: term()) :: term()

Called when the socket is reconnected.

This function takes the current state and returns the new state for the process.