Wumpex.Gateway (wumpex v0.1.0)

Handles incoming events from the gateway.

This module handles the raw incoming events from the gateway's Wumpex.Base.Websocket. The messages are decoded, events that are not specific to a guild, such as HELLO, IDENTIFY, INVALID_SESSION etc are handled inline. Events specific to a guild (MESSAGE_CREATE, ...) are passed on to the process group for that guild.

Types

frame_or_frames()

Specs

frame_or_frames() ::
  Wumpex.Base.Websocket.frame() | [Wumpex.Base.Websocket.frame()]

options()

Specs

options() :: [
  token: String.t(),
  shard: Wumpex.shard(),
  intents: Wumpex.Gateway.Intents.t(),
  handlers: [Wumpex.Gateway.Consumers.handler_options()]
]

The options that can be passed into start_link/1 and init/1.

Contains the following fields:

:token - The bot token

:shard - the identifier for this shard.

:intents - The Wumpex.Gateway.Intents.t/0 struct containing intent information.

:handlers - The handlers for this gateway's events.

state()

Specs

state() :: %{
  token: String.t(),
  ack: boolean(),
  sequence: non_neg_integer() | nil,
  session_id: String.t() | nil,
  shard: Wumpex.shard(),
  intents: non_neg_integer()
}

The state of the Wumpex.Gateway.Worker module.

:token - The bot token to authenticate against Discord.

:ack - Whether or not a heartbeat ACK has been received.

:sequence - The ID of the last received event.

:session_id - Session token, can be used to resume an interrupted session.

:shard - the identifier for this shard.

:intents - The intents information (calculated from Wumpex.Gateway.Intents.to_integer/1).

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

dispatch(event, state)

Specs

dispatch(event :: state(), state :: state()) :: state()

The dispatch event is called for all incoming events from the Discord gateway.

The Wumpex.Gateway.Worker tries to handle any non-specific event inline, maintains the sequence state and handles resumes. Unknown events are simply discarded and a warning is logged.

send_opcode(gateway, opcode)

Specs

send_opcode(gateway :: GenServer.server(), opcode :: map()) :: :ok

Dispatches an event on the gateway.

This method will handle rate limiting, encoding of the message etc. Note that there's no validation on what's being sent, so be careful invoking this method.

Invalid messages will cause Discord to close the gateway.

subscribe(shard)

Specs

subscribe(Wumpex.shard()) :: :ok

Subscribes the current process to events from the given gateway.

Events will be sent in the form of {:event, %Wumpex.Gateway.Event{}}.

subscribe(shard, pid_or_pids)

Specs

subscribe(Wumpex.shard(), pid_or_pids :: pid() | [pid()]) :: :ok

Subscribes the given processes to events from the given gateway.

Events will be sent in the form of {:event, %Wumpex.Gateway.Event{}}.

unsubscribe(shard)

Specs

unsubscribe(Wumpex.shard()) :: :ok | :not_joined

Unsubscribes the current process of events from the given gateway.

unsubscribe(shard, pid_or_pids)

Specs

unsubscribe(Wumpex.shard(), pid_or_pids :: pid() | [pid()]) :: :ok | :not_joined

Unsubscribes the given processes of events from the given gateway.

via(shard)

Specs

via(shard :: Wumpex.shard()) :: GenServer.server()

Gets the server name of the Wumpex.Gateway process for a given shard. You can use this to get the pid of the gateway, or to send opcodes to it.

# The shard (represented as a tuple).
{0, 1}
# Get the server name.
|> Wumpex.Gateway.via()
# Dispatch an opcode to it.
|> Wumpex.Gateway.send_opcode(%{"op" => 1})