Wumpex.Gateway.Caching (wumpex v0.1.0)

Handles state tracking as described in the official Discord documentation.

This is the second stage in the event processing pipeline, right after Wumpex.Gateway.EventProducer. Once an event is properly handled and the state is up to date, the event will be dispatched to all the listeners.

Types

options()

Specs

options() :: [{:producer, pid()}]

Represents the options to be passed to start_link/1.

Contains the following fields:

:producer - The pid/0 of the producer to subscribe to.