Wumpex.Gateway.Consumers (wumpex v0.1.0)

Handles starting new consumers for the event processing stages.

The start_consumer/2 method handles starting both Wumpex.Gateway.Consumers.EventConsumer as well as Wumpex.Gateway.Consumers.GuildsConsumer listeners.

Types

handler_options()

Specs

handler_options() :: [
  module: module(),
  filter: function(),
  guild: boolean() | non_neg_integer(),
  initial: [Wumpex.Gateway.Event.t()]
]

Represents the configuration for an event handler.

An event handler subscribes to events on the gateway, the :filter option can be used to subscribe to a limited number of events. The :guild option can be used to have an instance of the handler for each guild specific event.

Contains the following fields:

:module - The module that will receive the incoming events.

:filter - (optional) A function that takes in an event and returns true or false whether to accept this event or not

:guild - (optional) If set to true this handler will be instantiated for each guild and receive events for a specific guild.

:initial - (optional) A list of events that will be simulated on the handler upon startup.

Functions

start_consumer(producer, handler)

Specs

start_consumer(producer :: GenServer.server(), handler :: handler_options()) ::
  DynamicSupervisor.on_start_child()

Starts a new consumer subscribed to the given producer.

Depending on the given handler_options/0 either a event consumer or a guilds consumer will be started.