Wumpex.Gateway.Consumers.EventConsumer (wumpex v0.1.0)

Receives events from a Wumpex.Gateway.

The EventConsumer provides a convenience layer for subscribing to events from the event processing stages (see Wumpex.Gateway.EventProducer and Wumpex.Gateway.Caching). It implements GenStage as a :consumer and handles subscribing to the producer. It also takes care of filtering incoming events (through the :guild and :selector option) and allows specifying :initial events to replay when the listener starts up.

EventConsumers are generally started using Wumpex.Gateway.Consumers.start_consumer/2, but you can start one directly by calling start_link/1.

Types

options()

Specs

options() :: [
  producer: GenServer.server(),
  guild: String.t() | nil,
  handler: [
    module: module(),
    selector: function() | nil,
    initial: [Wumpex.Gateway.Event.t()] | nil
  ]
]

Represents the options that have to be passed into start_link/1.

Contains the following fields:

:producer - The producer to which the EventConsumer should subscribe to.

:guild - If not nil, all incoming events will be filtered to only include events from this guild.

:handler - Describes the handler for incoming events.

:module - The module that will receive incoming events.

:selector - If not nil, a function that receives an event and returns true or false to indicate if the event should be processed.

:initial - If not nil, a list of events that will be replayed on the handler upon startup.

state()

Specs

state() :: %{handler: module(), handler_state: term()}

Represents the process state.

Contains the following fields:

:handler - The module to which incoming events will be dispatched.

:handler_state - The state of the :handler module, generated from the :handler module's init function.

Functions

start_link(options)

Specs

start_link(options()) :: GenServer.on_start()

Start a new EventConsumer linked to the current process.