Wumpex.Gateway.Consumers.GuildsConsumer (wumpex v0.1.0)

A specialized event listener that checks for new guilds.

Like Wumpex.Gateway.Consumers.EventConsumer this module subscribes to Wumpex.Gateway.Caching and listens for incoming events. The main difference is that instead of forwarding incoming events to a handler, it starts a new EventConsumer scoped to a guild whenever a new guild becomes available.

This allows specifying a handler that should be instantiated for each guild that becomes available.

Types

guild_handler()

Specs

guild_handler() :: [module: module(), selector: function() | nil]

Represents the handler options for a handler that will be instantiated for each guild that becomes available.

Contains the following fields:

:module - The module to which events of a guild will be sent.

:selector - If not nil, a selector function to filter incoming events (see Wumpex.Gateway.Consumers.EventConsumer.options/0).

options()

Specs

options() :: [producer: GenServer.server(), guild_handler: guild_handler()]

Represents the options that can be passed into start_link/1.

Contains the following fields:

:producer - The producer to which the GuildsConsumer should subscribe to.

:guild_handler - The guild_handler/0 to instantiate for each guild.

state()

Specs

state() :: %{guild_handler: guild_handler(), producer: GenServer.server()}

Represents the process state.

Contains the following fields:

:guild_handler - The guild_handler/0 to instantiate for each guild.

:producer - The producer to which the EventConsumer should subscribe to.

Functions

start_link(options)

Specs

start_link(options()) :: GenServer.on_start()

Start a new GuildsConsumer linked to the current process.