Wumpex.Gateway.Event (wumpex v0.1.0)

Represents a single event dispatched from the Wumpex.Gateway.

Types

t()

Specs

t() :: %Wumpex.Gateway.Event{
  name: atom(),
  payload: map(),
  sequence: pos_integer(),
  shard: Wumpex.shard()
}

Represents an event that will be dispatched from the gateway to consumers.

Contains the following fields:

:shard - A Wumpex.shard/0 representing the shard from which the event originates.

:name - An atom with the name of the dispatched event.

:payload - The event payload in the form of a map.

:sequence - The sequence number of the event, can be used to track the same event across handlers.

Functions

guild(event)

Specs

guild(t()) :: pos_integer() | nil

Get the guild ID of a given event.

If the given event is not for a specific guild, this function will return nil.