Wumpex.Gateway.Intents (wumpex v0.1.0)

Contains method for representing and working with Gateway intents.

Types

t()

Specs

t() :: %Wumpex.Gateway.Intents{
  direct_message_reactions: boolean() | nil,
  direct_message_typing: boolean() | nil,
  direct_messages: boolean() | nil,
  guild_bans: boolean() | nil,
  guild_emojis: boolean() | nil,
  guild_integrations: boolean() | nil,
  guild_invites: boolean() | nil,
  guild_members: boolean() | nil,
  guild_message_reactions: boolean() | nil,
  guild_message_typing: boolean() | nil,
  guild_messages: boolean() | nil,
  guild_presences: boolean() | nil,
  guild_voice_states: boolean() | nil,
  guild_webhooks: boolean() | nil,
  guilds: boolean() | nil
}

Represents all possible intents and whether or not they should be enabled.

The full list of Discord intents can be found in the official Discord documentation.

Functions

privileged?(intents)

Specs

privileged?(intents :: t()) :: boolean()

Checks if this Intents struct contains privileged intents. Read more about privileged intents in the official Discord documentation.

iex> Wumpex.Gateway.Intents.privileged?(%Wumpex.Gateway.Intents{guilds: true})
false
iex> Wumpex.Gateway.Intents.privileged?(%Wumpex.Gateway.Intents{guild_members: true})
true

to_integer(intents)

Specs

to_integer(intents :: t()) :: non_neg_integer()

Transforms an Intents struct into the numerical representation.

iex> Wumpex.Gateway.Intents.to_integer(%Wumpex.Gateway.Intents{guilds: true})
1
iex> Wumpex.Gateway.Intents.to_integer(%Wumpex.Gateway.Intents{guilds: true, guild_messages: true})
513