Wumpex.Gateway.Opcodes (wumpex v0.1.0)

Provides methods for generating opcodes.

See the official documentation for a full list.

Types

opcode()

Specs

opcode() :: map()

Represents a generic OPCODE response.

Opcodes usually have a top level "op" field which contains the numerical ID and a "d" field which contains the payload.

Functions

heartbeat(sequence)

Specs

heartbeat(sequence :: non_neg_integer() | nil) :: opcode()

Generates a HEARTBEAT opcode.

iex> Wumpex.Gateway.Opcodes.heartbeat(nil)
%{"op" => 1, "d" => nil}

iex> Wumpex.Gateway.Opcodes.heartbeat(100)
%{"op" => 1, "d" => 100}

See the official documentation.

identify(token, shard, intents \\ 32767)

Specs

identify(
  token :: String.t(),
  shard :: Wumpex.shard(),
  intents :: non_neg_integer()
) :: opcode()

Generate an IDENTIFY opcode.

If you don't specify which intents you wish to use it will use 32767 by default, which uses ALL intents. Since version 8 this requires you to enable privileged gateway intents in the settings of your bot.

iex> Wumpex.Gateway.Opcodes.identify("test", {0, 1})
%{
  "op" => 2,
  "d" => %{
    "token" => "test",
    "shard" => [0, 1],
    "intents" => 32_767,
    "properties" => %{
      "$os" => :linux,
      "$browser" => "wumpex",
      "$device" => "wumpex"
    }
  }
}

iex> Wumpex.Gateway.Opcodes.identify("test", {1, 2}, 513)
%{
  "op" => 2,
  "d" => %{
    "token" => "test",
    "shard" => [1, 2],
    "intents" => 513,
    "properties" => %{
      "$os" => :linux,
      "$browser" => "wumpex",
      "$device" => "wumpex"
    }
  }
}

See the official documentation

resume(token, sequence, session_id)

Specs

resume(
  token :: String.t(),
  sequence :: non_neg_integer(),
  session_id :: String.t()
) :: opcode()

Generate a RESUME opcode.

iex> Wumpex.Gateway.Opcodes.resume("test", 20, "session")
%{
  "op" => 6,
  "d" => %{
    "token" => "test",
    "session_id" => "session",
    "seq" => 20
  }
}

See the official documentation

voice_state_update(guild, channel, options \\ [])

Specs

voice_state_update(
  guild :: String.t(),
  channel :: String.t() | nil,
  options :: keyword()
) :: opcode()