Wumpex.Resource (wumpex v0.1.0)

Base module for so called "resources" (as the discord documentation refers to them).

Resources are mainly used in the context of the HTTP API, but the gateway shares a lot of these types (so we use the same there).

Types

snowflake()

Specs

snowflake() :: String.t()

Represents a uniquely identifiable descriptor, guaranteed to be unique across all of Discord.

See the official Discord documentation.

Functions

to_atom!(value)

Specs

to_atom!(value :: String.t() | atom()) :: atom()

Attempts to convert a given string into an (existing!) atom/0.

This uses String.to_existing_atom/1 under the hood, so passing a non-existing atom will cause an ArgumentError

Examples

iex> Wumpex.Resource.to_atom!("known")
:known

iex> Wumpex.Resource.to_atom!("cannot-exist")
** (ArgumentError) argument error

to_atomized_map(value)

Specs

to_atomized_map(value :: map()) :: map()

Transforms a map with strings for keys into a map with atoms for keys.

Example

iex> Wumpex.Resource.to_atomized_map(%{"hello" => "world"})
%{hello: "world"}

to_datetime(value)

Specs

to_datetime(value :: DateTime.t() | nil | String.t()) :: DateTime.t() | nil

Attempts to convert a given string into a DateTime.t/0 instance.

Examples

iex> Wumpex.Resource.to_datetime("2017-07-11T17:27:07.299000+00:00")
~U[2017-07-11 17:27:07.299000Z]

iex> Wumpex.Resource.to_datetime(0)
~U[1970-01-01 00:00:00.000Z]

iex> Wumpex.Resource.to_datetime(~U[2020-07-11 17:27:07.299000Z])
~U[2020-07-11 17:27:07.299000Z]

to_structs(values, module)

Specs

to_structs(values :: [map()], module :: module()) :: [map()]

Maps a list of resources to the given module struct.

Example

iex> Wumpex.Resource.to_structs([%{}], Wumpex.Resource.Role)
[%Wumpex.Resource.Role{}]