Wumpex.Resource.Activity (wumpex v0.1.0)

Represents the activity of a presence update in Discord.

See the official Discord documentation.

Types

activity_type()

Specs

activity_type() :: 0 | 1 | 2 | 4

Represents the type of activity.

Can contain the following values:

ID	Name	Format	Example

0	Game	Playing {name}	"Playing Rocket League"

1	Streaming	Streaming {details}	"Streaming Rocket League"

2	Listening	Listening to {name}	"Listening to Spotify"

4	Custom	{emoji} {name}	":smiley: I am cool"

t()

Specs

t() :: %Wumpex.Resource.Activity{
  application_id: Wumpex.Resource.snowflake(),
  assets: Wumpex.Resource.Activity.Assets.t(),
  created_at: DateTime.t(),
  details: String.t(),
  emoji: Wumpex.Resource.Activity.Emoji.t(),
  flags: Wumpex.Resource.Activity.Flags.t(),
  instance: boolean(),
  name: String.t(),
  party: Wumpex.Resource.Activity.Party.t(),
  secrets: Wumpex.Resource.Activity.Secrets.t(),
  state: String.t(),
  timestamps: Wumpex.Resource.Activity.Timestamps.t(),
  type: activity_type(),
  url: String.t()
}

Represents the struct form of this module.

Contains the following fields:

:name - the activity's name

:type - the activity_type/0 activity type.

:url - stream URL, is validated when :type is 1.

:created_at - unix timestamp of when the activity was added to the user's session.

:timestamps - unix t:TWumpex.Resource.Timestamps.t/0 for start and end of the game.

:application_id - application id for the game.

:details - what the player is currently doing.

:state - the user's current party status.

:emoji - the emoji used for a custom status.

:party - information for the current party of the player.

:assets - images for the presence and their hover texts.

:secrets - secrets for Rich Presence joining and spectating.

:instance - whether or not the activity is an instance game session.

:flags - Wumpex.Resource.Activity.Flags.t/0 describes what the payload includes.

Functions

to_struct(data)

Specs

to_struct(data :: map()) :: t()

Maps the incoming data into struct form.

Example:

You can pass in invalid or missing data, it will ignore what doesn't match.

iex> Wumpex.Resource.Activity.to_struct(%{})
%Wumpex.Resource.Activity{
  application_id: nil,
  assets: nil,
  created_at: nil,
  details: nil,
  emoji: nil,
  flags: nil,
  instance: nil,
  name: nil,
  party: nil,
  secrets: nil,
  state: nil,
  timestamps: nil,
  type: nil,
  url: nil
}

If you pass in known properties, they'll be mapped.

iex> Wumpex.Resource.Activity.to_struct(%{"name" => "the name of the activity"})
%Wumpex.Resource.Activity{
  name: "the name of the activity"
}