Wumpex.Resource.Channel (wumpex v0.1.0)

Represents a guild or DM channel within Discord.

See the official Discord documentation.

Types

channel_type()

Specs

channel_type() :: 0 | 1 | 2 | 3 | 4 | 5 | 6

the type of channel

Can have the following values:

0 (GUILD_TEXT) - A text channel withing a server.

1 (DM) - A direct message between users.

2 (GUILD_VOICE) - A voice channel within a server.

3 (GROUP_DM) - A direct message between multiple users.

4 (GUILD_CATEGORY) - An organizational category that contains up to 50 channels.

5 (GUILD_NEWS) - A channel that users can follow and crosspost into their own server.

6 (GUILD_STORE) - A channel in which game developers can sell their game on Discord.

t()

Specs

t() :: %Wumpex.Resource.Channel{
  application_id: Wumpex.Resource.snowflake(),
  bitrate: non_neg_integer(),
  guild_id: Wumpex.Resource.snowflake(),
  icon: String.t(),
  id: Wumpex.Resource.snowflake(),
  last_message_id: Wumpex.Resource.snowflake(),
  last_pin_timestamp: DateTime.t(),
  name: String.t(),
  nsfw: boolean(),
  owner_id: Wumpex.Resource.snowflake(),
  parent_id: Wumpex.Resource.snowflake(),
  permission_overwrites: [Wumpex.Resource.Channel.Overwrite.t()],
  position: non_neg_integer(),
  rate_limit_per_user: non_neg_integer(),
  recipients: [Wumpex.Resource.User.t()],
  topic: String.t(),
  type: channel_type(),
  user_limit: non_neg_integer()
}

Represents the struct form of this module.

Contains the following fields:

:id - The ID of the channel.

:type - The channel_type/0 of the channel.

:guild_id - The ID of the guild.

:position - The sorting position of the channel.

:permission_overwrites - Explicit permission overwrites for members and roles.

:name - The name of the channel (2-100 characters).

:topic - The channel topic (0-1024 characters).

:nsfw - Whether the channel is NSFW.

:last_message_id - The ID of the last message sent in this channel (may not point to an existing or valid message).

:bitrate - The bitrate (in bits) of the voice channel.

:user_limit - The user limit of the voice channel.

:rate_limit_per_user - The amount of seconds a user has to wait before sending another message (between 0 and 21600); bots, users with the manage_messages or manage_channel are unaffected.

:recipients a list of Wumpex.Resource.User that are the recipients of the DM.

:icon The icon hash.

:owner_id - The ID of the DM creator.

:application_id - The application ID of the DM creator, if it's bot-created.

:parent_id - ID of the parent category for a channel (each parent category can contain up to 50 channels).

:last_pin_timestamp - When the last pinned message was pinned.

Functions

to_struct(data)

Specs

to_struct(data :: map()) :: t()

Maps the incoming data into struct form.

Example:

You can pass in invalid or missing data, it will ignore what doesn't match.

iex> Wumpex.Resource.Channel.to_struct(%{})
%Wumpex.Resource.Channel{
  id: nil,
  type: nil,
  guild_id: nil,
  position: nil,
  permission_overwrites: nil,
  name: nil,
  topic: nil,
  nsfw: nil,
  last_message_id: nil,
  bitrate: nil,
  user_limit: nil,
  rate_limit_per_user: nil,
  recipients: nil,
  icon: nil,
  owner_id: nil,
  application_id: nil
}