Wumpex.Resource.ClientStatus (wumpex v0.1.0)

Represents the status of a client on the three available platforms.

Active sessions are indicated with an "online", "idle", or "dnd" string per platform. If a user is offline or invisible, the corresponding field is not present.

See the official Discord documentation.

Types

status()

Specs

status() :: String.t()

Can be either "idle", "dnd", "online", or "offline"

t()

Specs

t() :: %Wumpex.Resource.ClientStatus{
  desktop: status(),
  mobile: status(),
  web: status()
}

Represents the struct form of this module.

Contains the following fields:

:desktop - The user's status set for an active desktop (Windows, Linux, Mac) application session.

:mobile - The user's status set for an active mobile (iOS, Android) application session.

:web - The user's status set for an active web (browser, bot account) application session.

Functions

to_struct(data)

Specs

to_struct(data :: map()) :: t()

Maps the incoming data into struct form.

Example:

You can pass in invalid or missing data, it will ignore what doesn't match.

iex> Wumpex.Resource.ClientStatus.to_struct(%{})
%Wumpex.Resource.ClientStatus{
  desktop: nil,
  mobile: nil,
  web: nil
}