Wumpex.Resource.Message (wumpex v0.1.0)

Represents a message sent in a channel within Discord.

See the official Discord documentation.

Types

message_type()

Specs

message_type() :: 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 14 | 15

The type of message.

Can have the following values:

0 - Regular message.

1 - Recipient was added.

2 - Recipient was removed.

3 - Call started.

4 - Channel name changed.

5 - Channel icon changed.

6 - Channel message was pinned.

7 - A guild member joined.

8 - User boosted the server.

9 - A user boosted the server to tier 1.

10 - A user boosted the server to tier 2.

11 - A user boosted the server to tier 3.

12 - A channel follow add.

14 - Guild discovery disqualified.

15 - Guild discovery re-qualified.

t()

Specs

t() :: %Wumpex.Resource.Message{
  activity: Wumpex.Resource.Message.Activity.t(),
  application: Wumpex.Resource.Message.Application.t(),
  attachments: [Wumpex.Resource.Attachment.t()],
  author: Wumpex.Resource.User.t(),
  channel_id: Wumpex.Resource.snowflake(),
  content: String.t(),
  edited_timestamp: DateTime.t(),
  embeds: [Wumpex.Resource.Embed.t()],
  flags: non_neg_integer(),
  guild_id: Wumpex.Resource.snowflake(),
  id: Wumpex.Resource.snowflake(),
  member: Wumpex.Resource.Guild.Member.t(),
  mention_channels: [Wumpex.Resource.ChannelMention.t()],
  mention_everyone: boolean(),
  mention_roles: [String.t()],
  mentions: [Wumpex.Resource.User.t()],
  message_reference: Wumpex.Resource.Message.Reference.t(),
  nonce: String.t(),
  pinned: boolean(),
  reactions: [Wumpex.Resource.Message.Reaction.t()],
  timestamp: DateTime.t(),
  tts: boolean(),
  type: message_type(),
  webhook_id: Wumpex.Resource.snowflake()
}

Represents the struct form of this module.

Contains the following fields:

:id - The ID of the message.

:channel_id - The ID of the channel the message was sent in.

:guild_id - The ID of the guild the message was sent in.

:author - The author of this message (not guaranteerd to be a valid user).

:member - Member properties for the message author.

:content - The contents of the message.

:timestamp - The timestamp when this message was sent.

:edited_timestamp - When this message was edited (set to nil if it's never been edited).

:tts - Whether this was a TTS message.

:mention_everyone - Whether this message mentions everyone.

:mentions - An array of Wumpex.Resource.User objects mentioned in this message.

:mention_roles - A list of role IDs specifically mentioned in this message.

:mention_channels - lis of Wumpex.Resource.ChannelMention specifically mentioned in this message.

:attachments - Any attached files.

:embeds - Any embedded content.

:reactions - A list of Wumpex.Resource.Message.Reaction reacted to this message.

:nonce - A unique nonce, can be used for validating the message was sent.

:pinned - Whether this message is pinned.

:webhook_id - The ID of the webhook, if it was generated by one.

:type - The message_type/0.

:activity - A Wumpex.Resource.Activity object, sent with rich presence related chat embeds.

:application - sent with rich presence related chat embeds.

:message_reference - Reference data sent with crossposted messages.

:flags - Message flags, describes extra features of the message.

Functions

to_struct(data)

Specs

to_struct(data :: map()) :: t()

Maps the incoming data into struct form.

Example:

You can pass in invalid or missing data, it will ignore what doesn't match.

iex> Wumpex.Resource.Message.to_struct(%{})
%Wumpex.Resource.Message{}