Wumpex.Resource.User (wumpex v0.1.0)

Users in Discord are generally considered the base entity. Users can spawn across the entire platform, be members of guilds, participate in text and voice chat, and much more. Users are separated by a distinction of "bot" vs "normal". Although they are similar, bot users are automated users that are "owned" by another user. Unlike normal users, bot users do not have a limitation on the number of Guilds they can be a part of.

See the official Discord documentation.

Types

premium_type()

Specs

premium_type() :: 0 | 1 | 2

Premium types denote the level of premium a user has.

Can have the following values:

0 None

1 Nitro Classic

2 Nitro

t()

Specs

t() :: %Wumpex.Resource.User{
  avatar: String.t(),
  bot: boolean(),
  discriminator: String.t(),
  email: String.t(),
  flags: Wumpex.Resource.UserFlags.t(),
  id: Wumpex.Resource.snowflake(),
  locale: String.t(),
  member: Wumpex.Resource.Guild.Member.t(),
  mfa_enabled: boolean(),
  premium_type: non_neg_integer(),
  public_flags: Wumpex.Resource.UserFlags.t(),
  system: boolean(),
  username: String.t(),
  verified: boolean()
}

Represents the struct form of this module.

Contains the following fields:

:id - The ID of the user.

:username - The user's username, not unique across the platform.

:discriminator - The user's 4 digit discord-tag.

:avatar - The user's avatar hash.

:bot - Whether the user belongs to an OAuth2 application.

:system - Whether the user is an official Discord system user (part of the urgent message system).

:mfa_enabled - Whether the user has two factor enabled on their account.

:locale- The user's preferred locale.

:verified - Whether the user's email has been verified.

:email - The user's email.

:flags - The flags on a user's account.

:premium_type - The type of Nitro subscription of the user.

:public_flags - The public flags on a user's account.

:member - Only passed when the user object was received through a Wumpex.Resource.Message object.

Functions

to_struct(data)

Specs

to_struct(data :: map()) :: t()

Maps the incoming data into struct form.

Example:

You can pass in invalid or missing data, it will ignore what doesn't match.

iex> Wumpex.Resource.User.to_struct(%{})
%Wumpex.Resource.User{}