Wumpex.Sharding (wumpex v0.1.0)

Manages the shards connected to the gateway.

This module will retrieve the required information from the Discord API and start shards accordingly.

Types

options()

Specs

options() :: [{:handlers, [Wumpex.Gateway.Consumers.handler_options()]}]

Represents the options that can be passed to start_link/1.

Contains the following fields:

:handlers - The event handlers that will be registered for each shard.

state()

Specs

state() :: %{
  url: String.t(),
  concurrency: non_neg_integer(),
  handlers: [module()]
}

Represents the state of the Sharding process.

Contains the following fields:

:url - The URL of the Discord gateway to connect to.

:concurrency - How many shards can be started concurrently.

:handlers - A list of modules that handle incoming events on the gateway.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(options \\ [])

Specs

start_link(options()) :: GenServer.on_start()