Wumpex.Voice (wumpex v0.1.0)

Contains functionality for connecting and interacting with voice connections.

Functions

change_channel(voice, channel)

Specs

change_channel(voice :: GenServer.server(), channel :: Wumpex.channel()) :: :ok

connect(shard, guild, channel)

Specs

connect(
  shard :: Wumpex.shard(),
  guild :: Wumpex.guild(),
  channel :: Wumpex.channel()
) :: Supervisor.on_start_child()

Connects to a given voice channel.

Returns {:ok, pid} where pid is the pid/0 of the voice connection. If there's already a voice connection active for the given guild, the bot will change to the given channel (using change_channel/2) and return {:ok, pid} as if started (allowing you to match on the same pattern for all success cases).

Takes the following parameters:

:shard - The shard on which the :guild is running.

:guild - The ID of the guild the given channel is in.

:channel - The ID of the channel to connect to.

disconnect(voice)

Specs

disconnect(voice :: GenServer.server()) :: :ok

Disconnects the given voice connection from it's currently connected channel.

for(guild)

Specs

for(guild :: Wumpex.guild()) :: GenServer.server()

Get the server name for the given guild.

play(voice, stream)

Specs

play(voice :: GenServer.server(), stream :: Enum.t()) :: reference()

Requests to play the given enumerable/stream on the given voice connection. The stream parameter can be both a list (or other enumerable) or a Stream.

This method will return a reference/0 that refers to this audio request. If you want to cancel the playback (either while it's playing or when it's still queued), you can use this reference to identity the playback request.

set_state(voice, options)

Specs

set_state(voice :: GenServer.server(), options :: [atom()]) :: :ok

Set the voice state for the given voice connection. This involves (un)muting and (un)deafening.

The :options parameter is a list that takes either :mute or :deafen as it's members. Passing in either of these two will cause the bot to mute or deafen (respectively).