Wumpex.Voice.Gateway (wumpex v0.1.0)

Handles incoming events from the voice gateway. The voice gateway is bound to a specific voice connection of a guild (see Wumpex.Voice.Manager).

This module handles the raw incoming events from the gateway's Wumpex.Base.Websocket. The messages are decoded and handled approperiately.

Types

frame_or_frames()

Specs

frame_or_frames() ::
  Wumpex.Base.Websocket.frame() | [Wumpex.Base.Websocket.frame()]

options()

Specs

options() :: [
  endpoint: String.t(),
  token: String.t(),
  session: String.t(),
  guild_id: String.t(),
  user_id: String.t(),
  controller: pid()
]

The options that can be passed into start_link/1 and init/1.

Contains the following fields:

:endpoint - The endpoint string received from the :VOICE_SERVER_UPDATE.

:token - The token received from the :VOICE_SERVER_UPDATE.

:session - The session token received from the :VOICE_STATE_UPDATE.

:guild_id - The guild ID to connect to.

:user_id - The ID of the bot.

:controller - The pid/0 of the controlling process (will receive the UDP information).

state()

Specs

state() :: %{
  nonce: non_neg_integer() | nil,
  controller: pid(),
  server: String.t(),
  session: String.t(),
  token: String.t()
}

Represents the process state.

Contains the following fields:

:nonce - The last nonce sent in the heartbeats.

:controller - The pid/0 of the controlling process (will receive the UDP information).

:server - The guild we're connected to

:session - The session token received from the :VOICE_STATE_UPDATE, used for resuming the connection.

:token - The token received from the :VOICE_SERVER_UPDATE, used for resuming the connection.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

select_protocol(gateway, ip, port, mode)

Specs

select_protocol(
  gateway :: pid(),
  ip :: String.t(),
  port :: non_neg_integer(),
  mode :: String.t()
) :: :ok

Asks the voice gateway to dispatch the SELECT PROTOCOL opcode with the given information.