Wumpex.Voice.Manager (wumpex v0.1.0)

Represents a voice connection and abstracts the Wumpex.Voice.Gateway, Wumpex.Voice.Udp connection and the Wumpex.Voice.Player process.

This module coordinates actions that require interaction across the different processes and/or modules. For example, sending audio data requires sending an opcode over the voice gateway and then sending encrypted audio data over the UDP socket.

Types

options()

Specs

options() :: [
  shard: Wumpex.shard(),
  guild: Wumpex.guild(),
  channel: Wumpex.channel()
]

The options that can be passed into start_link/1 and init/1.

Contains the following fields:

:shard - The shard on which to listen for the (regular) gateway events and send outgoing events.

:guild - The guild to connect to.

:channel - The (voice) channel to connect to.

state()

Specs

state() :: %{
  gateway: pid(),
  udp: pid(),
  player: pid(),
  ssrc: non_neg_integer(),
  shard: Wumpex.shard(),
  guild: Wumpex.guild(),
  channel: Wumpex.channel()
}

Represents the process state.

Contains the following fields:

:gateway - The pid/0 of the Wumpex.Voice.Gateway process.

:udp - The pid/0 of the Wumpex.Voice.Udp process.

:player - The pid/0 of the Wumpex.Voice.Player process.

:ssrc - The SSRC for this voice connection.

:shard - The shard on which to listen for the (regular) gateway events and send outgoing events.

:guild - The guild we're connected to.

:channel - The voice channel we're connected to.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

via(guild)

Specs

via(guild :: Wumpex.guild()) :: GenServer.server()

Gets the server name of the Wumpex.Voice.Manager process for a given guild. You can use this to get the pid of the connection or to send commands to it.

See also Wumpex.Gateway.via/1 for usage information.