Wumpex.Voice.Player (wumpex v0.1.0)

The Player module is responsible for encoding and sending audio data to Discord.

Currently only supports playing a single sound, but queueing support is planned to be added.

Types

options()

Specs

options() :: [secret_key: binary(), ssrc: non_neg_integer(), udp: pid()]

The options that can be passed into start_link/1 and init/1.

Contains the following fields:

:secret_key - The key used when encrypting outgoing packets.

:ssrc - The SSRC for this voice connection.

:udp - The pid/0 of the Wumpex.Voice.Udp process to request a socket from.

state()

Specs

state() :: %{
  ssrc: non_neg_integer(),
  secret_key: binary(),
  socket: Wumpex.Voice.Udp.socket(),
  current: reference() | nil,
  sequence: non_neg_integer(),
  time: non_neg_integer(),
  queue: [{reference(), Enum.t()}]
}

Represents the process state.

Contains the following fields:

:ssrc - The SSRC for this voice connection.

:secret_key - The key used when encrypting outgoing packets.

:socket - The Wumpex.Voice.Udp.socket/0 used for sending data.

:current - The currently being played audio data.

:sequence - The sequence used when building RTP header data.

:time - The time used when building RTP header data.

:queue - A list of all the audio fragments that are waiting to be played.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.