Wumpex.Voice.VoiceLedger (wumpex v0.1.0)

Keeps track of all voice connections by their name.

This module uses Wumpex.Base.Ledger under the hood.

Functions

lookup(name)

Specs

lookup(name :: any()) :: {pid(), any()} | nil

Looks up the pid/0 with metadata for a given name. Returns nil if no entry could be found for the given name.

iex> MyLedger.register("some-process", self(), "metadata")
:yes

iex> MyLedger.lookup("some-process")
{pid, "metadata"}

See register/3 for setting metadata.

register(name, pid, metadata \\ nil)

Specs

register(name :: any(), pid :: pid(), metadata :: any()) :: :yes | :no

Register a given pid/0 under the given :name in the registry, optionally passing in metadata.